Improve RPM script, add favicon and changes

+++ rpm/pre_uninstall.sh
@@ -1,2 +1,5 @@
-systemctl stop mcloud-importer
-systemctl disable mcloud-importer
+if [ "$1" = "0" ]; then
+  systemctl stop mcloud-importer
+  systemctl disable mcloud-importer
+fi
+exit 0
